# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EncadrePro API – gestion d'un atelier d'encadrement
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type DashboardStatsDevisEnAttente = {
  n: number;
  montant: number;
};

export type DashboardStatsFacturesImpayees = {
  n: number;
  montant: number;
};

export interface DashboardStats {
  caMois: number;
  devisEnAttente: DashboardStatsDevisEnAttente;
  facturesImpayees: DashboardStatsFacturesImpayees;
  nouveauxClients: number;
}

export interface CaMensuel {
  mois: string;
  total: number;
}

export interface RecentDoc {
  id: number;
  numero: string;
  client_nom: string;
  client_prenom?: string;
  total_ttc: number;
  statut: string;
  cree_le: string;
}

export interface Client {
  id: number;
  nom: string;
  prenom?: string | null;
  email?: string | null;
  telephone?: string | null;
  adresse?: string | null;
  ville?: string | null;
  code_postal?: string | null;
  notes?: string | null;
  cree_le: string;
  modifie_le: string;
}

export interface CreateClientBody {
  nom: string;
  prenom?: string | null;
  email?: string | null;
  telephone?: string | null;
  adresse?: string | null;
  ville?: string | null;
  code_postal?: string | null;
  notes?: string | null;
}

export interface Devis {
  id: number;
  numero: string;
  client_id: number;
  client_nom?: string | null;
  client_prenom?: string | null;
  date_creation: string;
  date_validite?: string | null;
  statut: string;
  sous_total_ht: number;
  total_tva_10: number;
  total_tva_20: number;
  total_ttc: number;
  notes?: string | null;
  conditions?: string | null;
  facture_id?: number | null;
  cree_le: string;
  modifie_le: string;
}

export interface Facture {
  id: number;
  numero: string;
  devis_id?: number | null;
  client_id: number;
  client_nom?: string | null;
  client_prenom?: string | null;
  date_creation: string;
  date_echeance?: string | null;
  statut: string;
  sous_total_ht: number;
  total_tva_10: number;
  total_tva_20: number;
  total_ttc: number;
  total_paye: number;
  solde_restant: number;
  notes?: string | null;
  conditions?: string | null;
  cree_le: string;
  modifie_le: string;
}

export interface ClientStats {
  devisCount: number;
  facturesCount: number;
  caTotal: number;
  devis: Devis[];
  factures: Facture[];
}

export interface Produit {
  id: number;
  reference?: string | null;
  designation: string;
  categorie: string;
  unite_calcul: string;
  prix_ht: number;
  taux_tva: number;
  actif: number;
  notes?: string | null;
  cree_le: string;
}

export interface CreateProduitBody {
  reference?: string | null;
  designation: string;
  categorie: string;
  unite_calcul: string;
  prix_ht: number;
  taux_tva: number;
  notes?: string | null;
}

export interface LigneDevis {
  id: number;
  devis_id: number;
  produit_id?: number | null;
  designation: string;
  unite_calcul: string;
  largeur_m?: number | null;
  hauteur_m?: number | null;
  quantite: number;
  quantite_calculee?: number | null;
  prix_unitaire_ht: number;
  taux_tva: number;
  total_ht: number;
  total_ttc: number;
  ordre: number;
}

export interface LigneInput {
  produit_id?: number | null;
  designation: string;
  unite_calcul: string;
  largeur_m?: number | null;
  hauteur_m?: number | null;
  quantite: number;
  prix_unitaire_ht: number;
  taux_tva: number;
  ordre?: number;
}

export type DevisDetail = Devis & {
  lignes: LigneDevis[];
};

export interface CreateDevisBody {
  client_id: number;
  notes?: string | null;
}

export interface SaveLignesBody {
  lignes: LigneInput[];
}

export interface UpdateStatutBody {
  statut: string;
}

export interface Acompte {
  id: number;
  facture_id: number;
  montant: number;
  date_paiement: string;
  mode_paiement?: string | null;
  notes?: string | null;
  cree_le: string;
}

export type FactureDetail = Facture & {
  lignes: LigneDevis[];
  paiements: Acompte[];
};

export interface CreateFactureBody {
  client_id: number;
  notes?: string | null;
}

export interface CreatePaiementBody {
  montant: number;
  date_paiement: string;
  mode_paiement?: string | null;
  notes?: string | null;
}

export interface Atelier {
  id: number;
  nom: string;
  siret?: string | null;
  adresse?: string | null;
  telephone?: string | null;
  email?: string | null;
  logo_path?: string | null;
  conditions_generales?: string | null;
  prefixe_devis: string;
  prefixe_facture: string;
  compteur_devis: number;
  compteur_facture: number;
  tva_defaut: number;
  email_template?: string | null;
  smtp_host?: string | null;
  smtp_port?: number | null;
  smtp_user?: string | null;
  smtp_pass?: string | null;
}

export interface SaveAtelierBody {
  nom?: string;
  siret?: string | null;
  adresse?: string | null;
  telephone?: string | null;
  email?: string | null;
  conditions_generales?: string | null;
  prefixe_devis?: string | null;
  prefixe_facture?: string | null;
  tva_defaut?: number | null;
  email_template?: string | null;
  smtp_host?: string | null;
  smtp_port?: number | null;
  smtp_user?: string | null;
  smtp_pass?: string | null;
}

export type ListClientsParams = {
  search?: string;
};

export type ListProduitsParams = {
  categorie?: string;
};

export type ListDevisParams = {
  statut?: string;
  client_id?: number;
};

export type ListFacturesParams = {
  statut?: string;
  client_id?: number;
};