# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: EncadrePro API – gestion d'un atelier d'encadrement
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"

    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /clients/{id}/stats:
    get:
      operationId: getClientStats
      tags: [clients]
      summary: Get stats for a specific client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientStats"

  # ──────────────── PRODUITS ────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── DASHBOARD ──
...
    Client:
      type: object
      properties:
        id: { type: integer }
        nom: { type: string }
        prenom: { type: string, nullable: true }
        email: { type: string, nullable: true }
        telephone: { type: string, nullable: true }
        adresse: { type: string, nullable: true }
        ville: { type: string, nullable: true }
        code_postal: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        cree_le: { type: string }
        modifie_le: { type: string }
      required: [id, nom, cree_le, modifie_le]
    CreateClientBody:
      type: object
      properties:
        nom: { type: string }
        prenom: { type: string, nullable: true }
        email: { type: string, nullable: true }
        telephone: { type: string, nullable: true }
        adresse: { type: string, nullable: true }
        ville: { type: string, nullable: true }
        code_postal: { type: string, nullable: true }
        notes: { type: string, nullable: true }
      required: [nom]
    ClientStats:
      type: object
      properties:
        devisCount: { type: integer }
        facturesCount: { type: integer }
        caTotal: { type: number }
        devis:
          type: array
          items:
            $ref: "#/components/schemas/Devis"
        factures:
          type: array
          items:
            $ref: "#/components/schemas/Facture"
      required: [devisCount, facturesCount, caTotal, devis, factures]

    # ── PRODUITS ──
...
    Devis:
      type: object
      properties:
        id: { type: integer }
        numero: { type: string }
        client_id: { type: integer }
        client_nom: { type: string, nullable: true }
        client_prenom: { type: string, nullable: true }
        date_creation: { type: string }
        date_validite: { type: string, nullable: true }
        statut: { type: string }
        sous_total_ht: { type: number }
        total_tva_10: { type: number }
        total_tva_20: { type: number }
        total_ttc: { type: number }
        notes: { type: string, nullable: true }
        conditions: { type: string, nullable: true }
        facture_id: { type: integer, nullable: true }
        cree_le: { type: string }
        modifie_le: { type: string }
      required: [id, numero, client_id, date_creation, statut, sous_total_ht, total_tva_10, total_tva_20, total_ttc, cree_le, modifie_le]
...
    Facture:
      type: object
      properties:
        id: { type: integer }
        numero: { type: string }
        devis_id: { type: integer, nullable: true }
        client_id: { type: integer }
        client_nom: { type: string, nullable: true }
        client_prenom: { type: string, nullable: true }
        date_creation: { type: string }
        date_echeance: { type: string, nullable: true }
        statut: { type: string }
        sous_total_ht: { type: number }
        total_tva_10: { type: number }
        total_tva_20: { type: number }
        total_ttc: { type: number }
        total_paye: { type: number }
        solde_restant: { type: number }
        notes: { type: string, nullable: true }
        conditions: { type: string, nullable: true }
        cree_le: { type: string }
        modifie_le: { type: string }
      required: [id, numero, client_id, date_creation, statut, sous_total_ht, total_tva_10, total_tva_20, total_ttc, total_paye, solde_restant, cree_le, modifie_le]
...