# lib/api-zod/src/generated/types/clientStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EncadrePro API – gestion d'un atelier d'encadrement
 * OpenAPI spec version: 1.0.0
 */
import type { Devis } from "./devis";
import type { Facture } from "./facture";

export interface ClientStats {
  devisCount: number;
  facturesCount: number;
  caTotal: number;
  devis: Devis[];
  factures: Facture[];
}